Fix failing tests since #1590 changes
