+++ tck/faces23/searchExpression/src/test/java/ee/jakarta/tck/faces/test/javaee8/searchExpression_selenium/Issue4331IT.java
@@ -17,7 +17,6 @@
 package ee.jakarta.tck.faces.test.javaee8.searchExpression_selenium;
 
 
-import static ee.jakarta.tck.faces.test.javaee8.searchExpression_selenium.Spec1238IT.findBehaviorScript;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
@@ -48,7 +47,7 @@ public class Issue4331IT extends BaseITNG {
 
         WebElement input = getWebDriver().findElement(By.id("issue4331ITInput1"));
 
-        String behaviorScript = findBehaviorScript(page, input);
+        String behaviorScript = getBehaviorScript(page, input);
 
         assertFalse(behaviorScript.contains("@custom"));
         assertTrue(behaviorScript.contains("issue4331ITInput1"));
