+++ tck/faces23/searchExpression/src/test/java/ee/jakarta/tck/faces/test/javaee8/searchExpression_selenium/Spec1238IT.java
@@ -53,7 +53,7 @@ public class Spec1238IT extends BaseITNG {
 
         assertEquals(label.getDomAttribute("for"), input.getDomAttribute("id"));
 
-        String behaviorScript = findBehaviorScript(page, input);
+        String behaviorScript = getBehaviorScript(page, input);
 
         if (behaviorScript.contains("@this")) {
             assertTrue(behaviorScript.contains("@this spec1238ITinput2"));
@@ -61,22 +61,5 @@ public class Spec1238IT extends BaseITNG {
         else {
             assertTrue(behaviorScript.contains("spec1238ITinput1 spec1238ITinput2"));
         }
-    }
-
-    public static String findBehaviorScript(WebPage page, WebElement input) {
-        var id = input.getAttribute("id");
-
-        for (var script : page.findElements(By.tagName("script"))) {
-            var src = script.getAttribute("src");
-            if (src == null || src.isEmpty()) {
-                var content = script.getDomProperty("textContent");
-
-                if (content.contains("'" + id + "'") || content.contains("\"" + id + "\"")) {
-                    return content;
-                }
-            }
-        }
-
-        return null;
     }
 }
